requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
readability-lxml==0.8.1
beautifulsoup4>=4.12.0
tiktoken==0.5.2
openai==1.12.0
chromadb==0.4.22
numpy>=1.24.0
responses==0.24.1

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0